layout/inspector/tests/test_color_to.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test InspectorUtils::ColorTo</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
  const InspectorUtils = SpecialPowers.InspectorUtils;

  const kEpsilon = 0.001;

  testColor("chocolate", "srgb", "rgb(210, 105, 30)", [0.8235294222831726, 0.4117647111415863, 0.11764705926179886, 1], false);

  testColor("chocolate", "hsl", "hsl(25 75% 47.0588%)", [25, 0.75, 0.4706], false);
  testColor("chocolate", "hwb", "hwb(25 11.7647% 17.6471%)", [25, 0.1176, 0.1765], false);
  testColor("chocolate", "lab", "lab(56.6293 39.2463 57.5567)", [56.6293, 39.2463, 57.5567], false);
  testColor("chocolate", "lch", "lch(56.6293 69.6638 55.711)", [56.6293, 69.6638, 55.711], false);
  testColor("chocolate", "oklab", "oklab(0.634398 0.099074 0.119193)", [0.634398, 0.099074, 0.119193], false);
  testColor("chocolate", "oklch", "oklch(0.634398 0.154992 50.2665)", [0.634398, 0.154992, 50.2665], false);

  testColor("chocolate", "srgb-linear", "color(srgb-linear 0.64448 0.141263 0.012983)", [0.64448, 0.141263, 0.012983], false);
  testColor("chocolate", "display-p3", "color(display-p3 0.770569 0.434015 0.199849)", [0.770569, 0.434015, 0.199849], false);
  testColor("chocolate", "a98-rgb", "color(a98-rgb 0.730405 0.410688 0.162005)", [0.730405, 0.410688, 0.162005], false);
  testColor("chocolate", "prophoto-rgb", "color(prophoto-rgb 0.592311 0.394149 0.164286)", [0.592311, 0.394149, 0.164286], false);
  testColor("chocolate", "rec2020", "color(rec2020 0.669266 0.401901 0.142716)", [0.669266, 0.401901, 0.142716], false);
  testColor("chocolate", "xyz-d50", "color(xyz-d50 0.337301 0.245449 0.0319589)", [0.337301, 0.245449, 0.0319589], false);
  testColor("chocolate", "xyz-d65", "color(xyz-d65 0.318634 0.239006 0.041637)", [0.318634, 0.239006, 0.041637], false);
  testColor("chocolate", "xyz", "color(xyz-d65 0.318634 0.239006 0.041637)", [0.318634, 0.239006, 0.041637], false);

  // Invalid color.
  testColor("darkblueorange", "srgb", null, null, null);
  testColor("rgb(none, none, none)", "lab", null, null, null);
  testColor(null, "lab", null, null, null);

  // Invalid color spaces.
  testColor("chocolate", "something", null, null, null);
  testColor("chocolate", "", null, null, null);
  testColor("chocolate", null, null, null, null);

  function testColor(color, colorSpace, expected, expectedComponents, expectedAdjust) {
    let actual = InspectorUtils.colorTo(color, colorSpace);

    // If we only get 3 elements for the expected components, we just add the
    // default alpha.
    if (expectedComponents && expectedComponents.length === 3) {
      expectedComponents.push(1);
    }

    if (actual === null) {
      is(expected, null, "color: converting " + color + " to " + colorSpace + " returns null");
      return;
    }

    is(actual.color, expected, "color is correctly converted to " + colorSpace);
    for (let i = 0; i < 4; i++) {
      isfuzzy(actual.components[i], expectedComponents[i], kEpsilon, "component " + i + " when converting to " + colorSpace);
    }
    is(actual.adjusted, expectedAdjust, "color adjusted status currect");
  }
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
